backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2
django-environ
django-channels
redis
django-redis
boto3
django-storages
Pillow
djoser
djangorestframework-simplejwt
django-cors-headers